+++ .env.example
@@ -15,6 +15,8 @@ NOTIFICATION_SERVICE_PG_SSL_ENABLED=false
 ## Integration With Amida Auth Microservice ##
 JWT_SECRET=0a6b944d-d2fb-46fc-a85e-0295c986cd9f
 PUSH_NOTIFICATIONS_SERVICE_USER_USERNAME=oucuYaiN6pha3ahphiiT
+PUSH_NOTIFICATIONS_SERVICE_USER_PASSWORD=T@estTest1
+AUTH_MICROSERVICE_URL=http://localhost:4000/api/v1
 
 ## Enabling Push Notifications ##
