+++ config/config.js
@@ -62,7 +62,9 @@ const config = {
     port: envVars.NOTIFICATION_SERVICE_PORT,
     jwtSecret: envVars.JWT_SECRET,
     automatedTestJwt: envVars.NOTIFICATION_SERVICE_AUTOMATED_TEST_JWT,
+    authMicroserviceUrl: envVars.AUTH_MICROSERVICE_URL,
     pushNotificationsServiceUserUsername: envVars.PUSH_NOTIFICATIONS_SERVICE_USER_USERNAME,
+    pushNotificationsServiceUserPassword: envVars.PUSH_NOTIFICATIONS_SERVICE_USER_PASSWORD,
     apnKeyId: envVars.PUSH_NOTIFICATIONS_APN_KEY_ID,
     apnTeamId: envVars.PUSH_NOTIFICATIONS_APN_TEAM_ID,
     apnEnv: envVars.PUSH_NOTIFICATIONS_APN_ENV,
