+++ package.json
@@ -11,7 +11,7 @@
     "yarn": ">=0.20.3"
   },
   "scripts": {
-    "start": "gulp serve",
+    "start": "yarn create-push-notifications-service-user && gulp serve",
     "create_db": "cross-env NODE_ENV=test node dist/create_db.js",
     "delete_db": "cross-env NODE_ENV=test node dist/delete_db.js",
     "start:debug": "cross-env DEBUG=amida-notification-microservice:* yarn start",
@@ -26,7 +26,7 @@
     "test:check-coverage": "yarn test:coverage && istanbul check-coverage",
     "report-coverage": "coveralls < ./coverage/lcov.info",
     "snyk-protect": "snyk protect",
-    "create-push-notifications-service-user": "node tools/createPushNotificationsServiceUser.js"
+    "create-push-notifications-service-user": "gulp && cross-env node dist/createPushNotificationsServiceUser.js"
   },
   "dependencies": {
     "bluebird": "3.4.6",
